Rename logcall to tracecall and split it out.
`logcall` has been split out into its own module, and renamed to
`tracecall`, as it's more accurately a tracing function, which traces
the calls and returns of the function. It's been split-out to ensure
that other modules can use it, portably, and to make it easier to test
ellis.py and the decorator. Ellis.py imports it as logcall for now.